# backend/services/ms-users/docs/swagger.yaml
basePath: /
definitions:
  apiresult.ErrResult:
    properties:
      error:
        $ref: '#/definitions/apiresult.ErrorPayload'
      ok:
        type: boolean
    type: object
  apiresult.ErrorPayload:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  finexa-ia_ms-users_internal_models.HealthResponse:
    properties:
      database:
        description: '"ok" when Postgres Ping works'
        example: ok
        type: string
      status:
        example: ok
        type: string
    type: object
  finexa-ia_ms-users_internal_models.UpsertUserRequest:
    properties:
      cognito_sub:
        example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        type: string
      email:
        example: jane@example.com
        type: string
    type: object
  finexa-ia_ms-users_internal_models.User:
    properties:
      cognito_sub:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  finexa-ia_ms-users_internal_models.UserOKResult:
    properties:
      data:
        $ref: '#/definitions/finexa-ia_ms-users_internal_models.User'
      ok:
        type: boolean
    type: object
info:
  contact: {}
  description: Usuarios internos (Cognito sub) en Postgres; misma BD que ms-plaid
    para FK.
  title: ms-users API
  version: 1.0.0
paths:
  /health:
    get:
      description: Comprueba que el proceso responde y que la base de datos acepta
        conexión
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finexa-ia_ms-users_internal_models.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/finexa-ia_ms-users_internal_models.HealthResponse'
      summary: Health check
      tags:
      - health
  /v1/users:
    post:
      consumes:
      - application/json
      description: Upsert por cognito_sub; email opcional
      parameters:
      - description: Cuerpo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/finexa-ia_ms-users_internal_models.UpsertUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finexa-ia_ms-users_internal_models.UserOKResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
      summary: Registrar o actualizar usuario
      tags:
      - users
  /v1/users/{id}:
    get:
      parameters:
      - description: ID interno
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finexa-ia_ms-users_internal_models.UserOKResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
      summary: Obtener usuario por id
      tags:
      - users
  /v1/users/by-cognito:
    get:
      parameters:
      - description: Sub del JWT Cognito
        in: query
        name: cognito_sub
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finexa-ia_ms-users_internal_models.UserOKResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresult.ErrResult'
      summary: Obtener usuario por Cognito sub
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"