# backend/services/ms-users/internal/repository/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlcgen

import (
	"database/sql"
	"time"
)

type User struct {
	ID         int64          `json:"id"`
	CognitoSub string         `json:"cognito_sub"`
	Email      sql.NullString `json:"email"`
	CreatedAt  time.Time      `json:"created_at"`
	UpdatedAt  time.Time      `json:"updated_at"`
	DeletedAt  sql.NullTime   `json:"deleted_at"`
}